refactor(test-harness): one type per file — split xlsx Sheet and RunFailures
output/xlsx/mod.rs held Sheet and Xlsx; Sheet moves to its own module (its ALL
and spec become pub since Xlsx now reaches them across the file). RunFailures
leaves benchmark/test/run.rs for benchmark/run_failures.rs — its lowest common
parent with Benchmark, which constructs it while Run only carries it.
